edited terraform and added typechecking test

+++ Clean-up/clean_lambda.py
@@ -23,11 +23,18 @@ def get_connection(config: _Environ) -> connection:
 
 def get_cursor(conn: connection) -> cursor:
     "Returns a cursor for the database"
+    if not isinstance(conn, connection):
+        raise TypeError(
+            'A cursor can only be constructed from a Psycopg2 connection object')
+
     return conn.cursor(cursor_factory=RealDictCursor)
 
 
 def delete_unsubscribed(conn: connection, table: str) -> list[dict]:
     "Deletes info about the unsubscribed products from the database"
+    if not isinstance(table, str):
+        raise TypeError('table must be of type string')
+
     with get_cursor(conn) as cur:
         cur.execute(f"""
             DELETE FROM {table}
